xlsx: utils.core.chart_to_xlsx
