+++ superset/viz.py
@@ -630,37 +630,11 @@ class BaseViz:
         return df.to_csv(index=include_index, **config["CSV_EXPORT"])
 
     def get_xlsx(self) -> Optional[bytes]:
-        from superset import db
-        from superset.models.slice import Slice
-
-        slice_id = int(self.form_data.get("slice_id"))
-        slice = db.session.query(Slice).filter(Slice.id == slice_id).one()
-
-
-        output = BytesIO()
-        writer = pd.ExcelWriter(output, engine="xlsxwriter")
-
-        image_data = self.form_data.get("image_data")
-        if image_data:
-            sheet = writer.book.add_worksheet("Sheet1")
-            sheet.write("A1", slice.slice_name)
-            image_data = image_data.split(",")[1]
-            image_data = base64.b64decode(image_data)
-            sheet.insert_image(
-                "A2", "in-memory", options={"image_data": BytesIO(image_data)}
-            )
-            del self.form_data["image_data"]
-        else:
-            df = self.get_df_payload()["df"]  # leverage caching logic
-            # Remove TZ from datetime64[ns, *] fields b4 writing to XLSX
-            df_clear_timezone(df)
-            include_index = not isinstance(df.index, pd.RangeIndex)
-            df.to_excel(writer, index=include_index, startrow=1)
-            sheet = writer.book.worksheets()[0]
-            sheet.write("A1", slice.slice_name)
-        writer.close()
-        output.seek(0)
-        return output.read()
+        return utils.chart_to_xlsx(
+            self.get_df_payload()["df"],
+            self.form_data.get("image_data"),
+            self.form_data.get("slice_id")
+        ).read()
 
     def get_data(self, df: pd.DataFrame) -> VizData:
         return df.to_dict(orient="records")
